implemented non-static method execution via field injection

+++ EpicAnnotations/src/main/java/edu/vanderbilt/clinicalsystems/epic/api/oo/ValueArray.java
@@ -33,7 +33,7 @@ public class ValueArray {
 	// NativeValueType.EXTENSION causes the method return value to be treated as an expression
 	// produces tools().expression().generate( "1" ) ;
 	@NativeValue(RESULT)
-	public int firstLine() { return 1 ; }
+	public long firstLine() { return 1 ; }
 	
 	// NativeValueType.RESULT causes the return value to be treated as an expression
 	// can only reference final fields
@@ -42,12 +42,12 @@ public class ValueArray {
 	// "m_delegate" is local variable assigned to "private final Value m_delegate"
 	// local variable assigned to "private final Value m_delegate" is the argument to "new ValueArray( ??? )"
 	@NativeValue(RESULT)
-	public int lastLine() { return m_delegate.get("0").toInt() ; }
+	public long lastLine() { return m_delegate.get("0").toLong() ; }
 	
 	// NativeComandType.EXTENSION causes the method statements to be treated as commands
 	@NativeCommand(EXTENSION)
 	public void addLine( Value value ) {
-		int nextLine = lastLine() + 1 ;
+		long nextLine = lastLine() + 1 ;
 		m_delegate.get("0").assign(String.valueOf(nextLine)) ;
 		m_delegate.get(String.valueOf(nextLine)).assign(value.toString()) ;
 	}
