+++ MLanguageModel/src/main/java/edu/vanderbilt/clinicalsystems/m/lang/model/argument/ColumnCommand.java
@@ -6,12 +6,12 @@ public class ColumnCommand extends FormatCommand {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private final int m_column ;
+	private final long m_column ;
 	
-	public ColumnCommand(int column) { m_column = column ; }
+	public ColumnCommand(long column) { m_column = column ; }
 	
-	@Override public String text() { return "?" + Integer.toString(m_column); }
-	public int column() { return m_column ; }
+	@Override public String text() { return "?" + Long.toString(m_column); }
+	public long column() { return m_column ; }
 	
 	@Override public <R> R visit( Visitor<R> visitor ) { return visitor.visitColumnCommand(this) ; }
 }
